pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pdfreading-app"
version = "0.1.7"
description = "A PDF reader desktop app"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pillow",
    "pymupdf",
    "customtkinter"
]

[project.scripts]
pdfreading-app = "pdfreading.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/pdfreading"]

[tool.hatch.build.targets.wheel.shared-data]
"src/pdfreading/assets" = "pdfreading/assets"